requirements.txt
markdown>=3.3.0
pyyaml>=6.0
pygments>=2.12.0
python-dateutil>=2.8.2
